Products/pyproject.toml
[tool.poetry]
name = "products"
version = "0.1.0"
description = ""
authors = ["Harmit <harmit.singh@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
flask-cors = "^3.0.10"
flask-sqlalchemy = "^3.0.3"
pyjwt = "^2.6.0"
marshmallow = "^3.19.0"
flask-jwt-extended = "^4.4.4"
marshmallow-sqlalchemy = "^0.29.0"
python-dotenv = "^1.0.0"
pymysql = "^1.0.2"
coverage = "^7.2.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
